refactor: Moving tests around to match Preact's style
