+++ package.json
@@ -15,7 +15,6 @@
     "preact-router": "^3.1.0"
   },
   "devDependencies": {
-    "@types/enzyme": "^3.10.3",
     "@types/history": "^4.7.3",
     "@types/jest": "^25.1.4",
     "@types/jsdom": "^16.1.0",
@@ -69,10 +68,10 @@
     "preset": "ts-jest",
     "testEnvironment": "node",
     "testPathIgnorePatterns": [
-      "<rootDir>/dist"
+      "<rootDir>/{dist,src}"
     ],
     "setupFilesAfterEnv": [
-      "<rootDir>/src/setup-test.ts"
+      "<rootDir>/tests/setup-test.ts"
     ],
     "collectCoverageFrom": [
       "<rootDir>/src/**/*.{ts,tsx}",
