+++ tests/components/ArticleCommentCard.spec.tsx
@@ -1,11 +1,12 @@
-import render from 'preact-render-to-string'
-import ArticleCommentCard from '../ArticleCommentCard'
 import { h } from 'preact'
+import render from 'preact-render-to-string'
 import { shallow } from 'enzyme'
-import { generateProfile, generateComments } from '../../utils/test-utils'
-import { useRootState } from '../../store'
 
-jest.mock('../../store')
+import ArticleCommentCard from '../../src/components/ArticleCommentCard'
+import { generateProfile, generateComments } from '../utils/test-utils'
+import { useRootState } from '../../src/store'
+
+jest.mock('../../src/store')
 
 const useRootStateMock = useRootState as jest.Mock
