+++ tests/components/Header.spec.tsx
@@ -1,9 +1,10 @@
 import { h } from 'preact'
-import Header from '../Header'
 import render from 'preact-render-to-string'
-import { useRootState } from '../../store'
 
-jest.mock('../../store')
+import Header from '../../src/components/Header'
+import { useRootState } from '../../src/store'
+
+jest.mock('../../src/store')
 
 const useRootStateMock = useRootState as jest.Mock
