+++ tests/components/Pagination.spec.tsx
@@ -1,7 +1,8 @@
-import render from 'preact-render-to-string'
-import Pagination from '../Pagination'
 import { h } from 'preact'
+import render from 'preact-render-to-string'
 import { shallow } from 'enzyme'
+
+import Pagination from '../../src/components/Pagination'
 
 describe('# Pagination Component', function () {
   const setPage = jest.fn()
