+++ tests/pages/Settings.spec.tsx
@@ -1,15 +1,16 @@
-import { mount, shallow } from 'enzyme'
 import { h } from 'preact'
-import Settings from '../Settings'
-import { useRootState } from '../../store'
 import { route } from 'preact-router'
-import { getInputValue, setInputValue } from '../../utils/test-utils'
-import { putProfile } from '../../services'
-import { UPDATE_USER } from '../../store/constants'
+import { mount, shallow } from 'enzyme'
 
-jest.mock('../../store')
+import Settings from '../../src/pages/Settings'
+import { useRootState } from '../../src/store'
+import { getInputValue, setInputValue } from '../utils/test-utils'
+import { putProfile } from '../../src/services'
+import { UPDATE_USER } from '../../src/store/constants'
+
+jest.mock('../../src/store')
 jest.mock('preact-router')
-jest.mock('../../services')
+jest.mock('../../src/services')
 
 const useRootStateMock = useRootState as jest.Mock
 const putProfileMock = putProfile as jest.Mock
