+++ tests/store/action.spec.ts
@@ -1,8 +1,8 @@
-import { postLogin } from '../../services'
-import { login } from '../actions'
-import { SET_ERRORS, UPDATE_USER } from '../constants'
+import { postLogin } from '../../src/services'
+import { login } from '../../src/store/actions'
+import { SET_ERRORS, UPDATE_USER } from '../../src/store/constants'
 
-jest.mock('../../services')
+jest.mock('../../src/services')
 
 const postLoginMock = postLogin as jest.Mock
